+++ packages/eslint-plugin-cds/src/rules/control-has-associated-label-extended.ts
@@ -25,7 +25,9 @@ type MessageIds =
   | 'missingCalendarIconButtonAccessibilityLabel'
   | 'missingNextArrowAccessibilityLabel'
   | 'missingPreviousArrowAccessibilityLabel'
-  | 'missingCardDismissAccessibilityLabel';
+  | 'missingCardDismissAccessibilityLabel'
+  | 'missingStartIconAccessibilityLabel'
+  | 'missingClearIconAccessibilityLabel';
 
 type ConditionalCheckType = {
   configArray: string[];
@@ -57,6 +59,7 @@ const config = {
   checkForCalendarIconButtonAccessibilityLabelProps: ['DatePicker'],
   checkForArrowAccessibilityProps: ['DatePicker', 'Calendar', 'TabNavigation'],
   checkForCardDismissAccessibilityLabelProps: ['NudgeCard', 'UpsellCard'],
+  checkForSearchInputAccessibilityLabelProps: ['SearchInput'],
 
   allowedPackages: ['@cbhq/cds-common', '@cbhq/cds-web', '@cbhq/cds-web-visualization'],
 };
@@ -80,6 +83,8 @@ export const controlHasAssociatedLabelExtended: TSESLint.RuleModule<MessageIds> 
       missingNextArrowAccessibilityLabel: `Missing 'nextArrowAccessibilityLabel' on <{{componentName}}>.`,
       missingPreviousArrowAccessibilityLabel: `Missing 'previousArrowAccessibilityLabel' on <{{componentName}}>.`,
       missingCardDismissAccessibilityLabel: `Missing 'accessibilityLabel' on <{{componentName}}> for dismiss button.`,
+      missingStartIconAccessibilityLabel: `Missing 'startIconAccessibilityLabel' on <{{componentName}}>.`,
+      missingClearIconAccessibilityLabel: `Missing 'clearIconAccessibilityLabel' on <{{componentName}}>.`,
     },
     hasSuggestions: true,
     fixable: 'code',
@@ -124,6 +129,8 @@ export const controlHasAssociatedLabelExtended: TSESLint.RuleModule<MessageIds> 
           hasMissingNextArrowAccessibilityLabel,
           hasMissingPreviousArrowAccessibilityLabel,
           hasOnDismissPressProp,
+          hasMissingStartIconAccessibilityLabel,
+          hasMissingClearIconAccessibilityLabel,
         } = extractA11yAttributesState(node, node.openingElement);
 
         let isTextInputWithNegativeVariant = true;
@@ -192,6 +199,18 @@ export const controlHasAssociatedLabelExtended: TSESLint.RuleModule<MessageIds> 
             condition: !hasAccessibilityLabel && hasOnDismissPressProp,
             messageId: 'missingCardDismissAccessibilityLabel',
             suggestedPropToAdd: 'accessibilityLabel',
+          },
+          {
+            configArray: config.checkForSearchInputAccessibilityLabelProps,
+            condition: !hasMissingStartIconAccessibilityLabel,
+            messageId: 'missingStartIconAccessibilityLabel',
+            suggestedPropToAdd: 'startIconAccessibilityLabel',
+          },
+          {
+            configArray: config.checkForSearchInputAccessibilityLabelProps,
+            condition: !hasMissingClearIconAccessibilityLabel,
+            messageId: 'missingClearIconAccessibilityLabel',
+            suggestedPropToAdd: 'clearIconAccessibilityLabel',
           },
         ];