+++ packages/eslint-plugin-cds/src/rules/has-valid-accessibility-descriptors-extended.ts
@@ -18,7 +18,9 @@ type MessageIds =
   | 'missingHandleBarAccessibilityLabel'
   | 'missingHelperTextErrorIconAccessibilityLabel'
   | 'missingCalendarIconButtonAccessibilityLabel'
-  | 'missingCardDismissAccessibilityLabel';
+  | 'missingCardDismissAccessibilityLabel'
+  | 'missingStartIconAccessibilityLabel'
+  | 'missingClearIconAccessibilityLabel';
 
 type ConditionalCheckType = {
   configArray: string[];
@@ -49,6 +51,7 @@ const config = {
   checkForHelperTextErrorIconAccessibilityLabelProps: ['TextInput'],
   checkForCalendarIconButtonAccessibilityLabelProps: ['DatePicker'],
   checkForCardDismissAccessibilityLabelProps: ['NudgeCard', 'UpsellCard'],
+  checkForSearchInputAccessibilityLabelProps: ['SearchInput'],
 
   allowedPackages: ['@cbhq/cds-common', '@cbhq/cds-mobile', '@cbhq/cds-mobile-visualization'],
 };
@@ -68,6 +71,8 @@ export const hasValidA11yDescriptorsExtended: TSESLint.RuleModule<MessageIds> = 
       missingHelperTextErrorIconAccessibilityLabel: `Missing 'helperTextErrorIconAccessibilityLabel' on <{{componentName}}>.`,
       missingCalendarIconButtonAccessibilityLabel: `Missing 'calendarIconButtonAccessibilityLabel' on <{{componentName}}>.`,
       missingCardDismissAccessibilityLabel: `Missing 'accessibilityLabel' on <{{componentName}}> for dismiss button.`,
+      missingStartIconAccessibilityLabel: `Missing 'startIconAccessibilityLabel' on <{{componentName}}>.`,
+      missingClearIconAccessibilityLabel: `Missing 'clearIconAccessibilityLabel' on <{{componentName}}>.`,
     },
     hasSuggestions: true,
     fixable: 'code',
@@ -110,6 +115,8 @@ export const hasValidA11yDescriptorsExtended: TSESLint.RuleModule<MessageIds> = 
           hasHelperTextErrorIconAccessibilityLabel,
           hasCalendarIconButtonAccessibilityLabel,
           hasOnDismissPressProp,
+          hasMissingStartIconAccessibilityLabel,
+          hasMissingClearIconAccessibilityLabel,
         } = extractA11yAttributesState(node, node.openingElement);
 
         let isTextInputWithNegativeVariant = true;
@@ -159,6 +166,16 @@ export const hasValidA11yDescriptorsExtended: TSESLint.RuleModule<MessageIds> = 
             configArray: config.checkForCardDismissAccessibilityLabelProps,
             condition: !hasAccessibilityLabel && hasOnDismissPressProp,
             messageId: 'missingCardDismissAccessibilityLabel',
+          },
+          {
+            configArray: config.checkForSearchInputAccessibilityLabelProps,
+            condition: !hasMissingStartIconAccessibilityLabel,
+            messageId: 'missingStartIconAccessibilityLabel',
+          },
+          {
+            configArray: config.checkForSearchInputAccessibilityLabelProps,
+            condition: !hasMissingClearIconAccessibilityLabel,
+            messageId: 'missingClearIconAccessibilityLabel',
           },
         ];